Move debugging/coreclr to src
