+++ extension.bundle.ts
@@ -26,18 +26,18 @@ export { throwDockerConnectionError, internal } from './explorer/utils/dockerCon
 export { getImageLabel } from './explorer/models/getImageLabel';
 export { trimWithElipsis } from './explorer/utils/utils';
 export { isWindows10RS3OrNewer, isWindows10RS4OrNewer, isWindows10RS5OrNewer } from "./src/utils/osVersion";
-export { LineSplitter } from './debugging/coreclr/lineSplitter';
-export { CommandLineBuilder } from './debugging/coreclr/commandLineBuilder';
-export { DockerClient } from './debugging/coreclr/dockerClient';
-export { LaunchOptions } from './debugging/coreclr/dockerManager';
-export { DotNetClient } from './debugging/coreclr/dotNetClient';
-export { FileSystemProvider } from './debugging/coreclr/fsProvider';
-export { OSProvider } from './debugging/coreclr/osProvider';
-export { DockerDaemonIsLinuxPrerequisite, DockerfileExistsPrerequisite, DotNetSdkInstalledPrerequisite, LinuxUserInDockerGroupPrerequisite, MacNuGetFallbackFolderSharedPrerequisite } from './debugging/coreclr/prereqManager';
-export { ProcessProvider } from './debugging/coreclr/processProvider';
+export { LineSplitter } from './src/debugging/coreclr/lineSplitter';
+export { CommandLineBuilder } from './src/debugging/coreclr/commandLineBuilder';
+export { DockerClient } from './src/debugging/coreclr/dockerClient';
+export { LaunchOptions } from './src/debugging/coreclr/dockerManager';
+export { DotNetClient } from './src/debugging/coreclr/dotNetClient';
+export { FileSystemProvider } from './src/debugging/coreclr/fsProvider';
+export { OSProvider } from './src/debugging/coreclr/osProvider';
+export { DockerDaemonIsLinuxPrerequisite, DockerfileExistsPrerequisite, DotNetSdkInstalledPrerequisite, LinuxUserInDockerGroupPrerequisite, MacNuGetFallbackFolderSharedPrerequisite } from './src/debugging/coreclr/prereqManager';
+export { ProcessProvider } from './src/debugging/coreclr/processProvider';
 export { PlatformOS, Platform } from './src/utils/platform';
-export { DockerBuildImageOptions } from "./debugging/coreclr/dockerClient";
-export { compareBuildImageOptions } from "./debugging/coreclr/dockerManager";
+export { DockerBuildImageOptions } from "./src/debugging/coreclr/dockerClient";
+export { compareBuildImageOptions } from "./src/debugging/coreclr/dockerManager";
 export { configure, ConfigureApiOptions, ConfigureTelemetryProperties } from './configureWorkspace/configure';
 export { globAsync } from './src/utils/async';
 export { httpsRequestBinary } from './src/utils/httpRequest';
