+++ src/debugging/coreclr/appStorage.ts
