+++ src/debugging/coreclr/debugSessionManager.ts
