+++ src/debugging/coreclr/fsProvider.ts
