+++ src/debugging/coreclr/lazy.ts
