+++ src/debugging/coreclr/netCoreProjectProvider.ts
