+++ src/debugging/coreclr/outputManager.ts
