+++ src/debugging/coreclr/prereqManager.ts
