+++ src/debugging/coreclr/tempFileProvider.ts
