+++ src/extension.ts
@@ -41,7 +41,6 @@ import { docker } from '../commands/utils/docker-endpoint';
 import { DefaultTerminalProvider } from '../commands/utils/TerminalProvider';
 import { DockerDebugConfigProvider } from '../configureWorkspace/configDebugProvider';
 import { configure, configureApi } from '../configureWorkspace/configure';
-import { registerDebugConfigurationProvider } from '../debugging/coreclr/registerDebugger';
 import { AzureAccountWrapper } from '../explorer/deploy/azureAccountWrapper';
 import { WebAppCreator } from '../explorer/deploy/webAppCreator';
 import { DockerExplorerProvider } from '../explorer/dockerExplorerProvider';
@@ -52,6 +51,7 @@ import { NodeBase } from '../explorer/models/nodeBase';
 import { browseAzurePortal } from '../explorer/utils/browseAzurePortal';
 import { browseDockerHub, dockerHubLogin, dockerHubLogout } from '../explorer/utils/dockerHubUtils';
 import { COMPOSE_FILE_GLOB_PATTERN, configPrefix, configurationKeys } from './constants';
+import { registerDebugConfigurationProvider } from './debugging/coreclr/registerDebugger';
 import { DockerComposeCompletionItemProvider } from './dockerCompose/dockerComposeCompletionItemProvider';
 import { DockerComposeHoverProvider } from './dockerCompose/dockerComposeHoverProvider';
 import composeVersionKeys from './dockerCompose/dockerComposeKeyInfo';
